+++ crates/coeus-ops/src/backend_ops/cpu_impl/impls/elementwise.rs
@@ -1,8 +1,8 @@
 use super::super::elementwise;
 use super::super::CpuBackend;
 use crate::backend_ops::ops::{BinaryOp, UnaryOp};
-use crate::backend_ops::traits::ElementwiseOps;
-use coeus_core::{CpuAddressableStorageMut, Layout, Scalar};
+use crate::backend_ops::traits::{ElementwiseOps, ScalarPowerOps};
+use coeus_core::{CpuAddressableStorageMut, Float, Layout, Scalar};
 
 #[allow(clippy::too_many_arguments)]
 impl<T: Scalar + leto_ops::Scalar, B: CpuBackend> ElementwiseOps<T> for B
@@ -64,5 +64,29 @@ where
         c_layout: &Layout,
     ) -> Result<(), Self::Error> {
         elementwise::elementwise_unary(self, op, a, a_layout, c, c_layout)
+    }
+}
+
+impl<T: Float + leto_ops::Scalar + leto_ops::RealScalar, B: CpuBackend> ScalarPowerOps<T> for B
+where
+    B::DeviceBuffer<T>: CpuAddressableStorageMut<T>,
+{
+    #[inline]
+    fn elementwise_pow_scalar(
+        &self,
+        input: &Self::DeviceBuffer<T>,
+        input_layout: &Layout,
+        exponent: T,
+        output: &mut Self::DeviceBuffer<T>,
+        output_layout: &Layout,
+    ) -> Result<(), Self::Error> {
+        elementwise::elementwise_pow_scalar(
+            self,
+            exponent,
+            input,
+            input_layout,
+            output,
+            output_layout,
+        )
     }
 }